Receiver: Refactor main file
Now we connect to a database that stores all keys, the file name, and file path. When needed, the file is decrypted so the frontend can show a preview/download the file as desired.

Also add support for basic video, audio and image file under 15 MB, these are stored on disk to not clutter the database.